+++ src/src/state.h
@@ -19,7 +19,7 @@ extern const char* vsn;
 void do_state();
 void init_state();
 //common variables used by multipe things
-extern volatile int rpm;
+extern volatile float rpm;
 extern double mmPerStep;
 extern int32_t stepsPerMM;
 extern int32_t relativePosition;
